Changing spring configuration file to use JPA instead of hibernate. Adding code for verification document functionality.

+++ src/main/java/com/bank/model/BaseEntity.java
@@ -15,14 +15,14 @@ import javax.persistence.MappedSuperclass;
 public class BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    protected Integer id;
+    protected Long id;
 
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
